Fix negative filters in datadog agent data check
If the filter string has a tag that starts with an exclamation mark.
These can not be used literally in the over part of a service check
query. We now filter those, no longer add them to the over part of the
query, but add them to the exclude part of the query.